backend/requirements/development.txt
-r base.txt
django-debug-toolbar>=4.2
django-extensions>=3.2
djangorestframework-simplejwt[crypto]>=5.3
